// src/contract/model/bookingDetailsDtoListApiResponse.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BookingDetailsDto } from './bookingDetailsDto';

export interface BookingDetailsDtoListApiResponse {
  isSuccess?: boolean;
  data?: Array<BookingDetailsDto> | null;
  message?: string | null;
}