// src/contract/model/rentalDetailsDtoListApiResponse.ts
/**
 * EV Rental API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RentalDetailsDto } from './rentalDetailsDto';

export interface RentalDetailsDtoListApiResponse {
  isSuccess?: boolean;
  data?: Array<RentalDetailsDto> | null;
  message?: string | null;
}